+++ stream/FixedMemoryStream.h
@@ -20,9 +20,9 @@ namespace CommonLib
 		~CFxMemoryWriteStream();
 			
 
-		virtual void WriteBytes(const byte_t* buffer, uint32_t size);
-		virtual void WriteInverse(const byte_t* buffer, uint32_t size);
-		virtual void WriteStream(IStream *pStream, int32_t nPos = -1, int32_t nSize = -1);
+		virtual void WriteBytes(const byte_t* buffer, size_t size);
+		virtual void WriteInverse(const byte_t* buffer, size_t size);
+		virtual void WriteStream(IStream *pStream, int64_t nPos = -1, int64_t nSize = -1);
 
 	};
 }
